data-collector/requirements.txt
requests==2.31.0
kafka-python==2.0.2
elasticsearch==8.11.0
python-dotenv==1.0.0
